- fix sign errors in rsubhi and divmodhi4 - added mul/div/mod unit test with a comprehensive list of values - added escape character fix to app.s in assembler - added some missing files including eqdf2.c and c-cppbuiltin.c

+++ test/harness/harness.cc
@@ -92,7 +92,7 @@ void Harness::test_report ()
 void Harness::test_printf()
 {
     int c;
-    int s[10];
+    short s[10];
     int stack;
 
     stack = memReadW (getWP() + 20); // R10
